requirements.txt
kafka-python
python-dotenv
pymysql
pandas
mysql-connector-python